Bugfix where the Toolkit could delete manual SP:s
If a SP has a name that doesn't correspond to what ADFS Toolkit would have given it (the domain part of the entityID) and the name don't start with the chosen prefix, it would be deleted ad re-added by the toolkit. This is now fixed and no SP without the prefix should be deleted. 

Also fixed a small bug where the separator (:) is missing for SP's that share the same domain name.